requirements.txt
h5py==3.7.0
matplotlib==3.5.1
numpy==1.21.5
pandas==1.4.2
python-dotenv==0.20.0
pytorch_lightning==1.6.3
PyWavelets==1.3.0
scikit_learn==1.1.2
scipy==1.7.3
torch==1.12.1
torchmetrics==0.8.1
torchvision==0.13.1
